Move `wipe` command into its own submodule similarly to other commands
